Rewrote the code-example
Rewrote the code-example, to fit with what was discussed during the meeting on 2/4.

RecipeBook calls on RecipeReader, which in turn reads all files in the "files"-directory, then creates a Recipe-object per file, and returns an ArrayList of the created objects.

RecipeBook then goes through all recipes and adds all ingredients to an ArrayList, skipping duplicates.